languages/Python/flask/requirements.txt
pyyaml
werkzeug
# sudo apt-get install libmagic1
python-magic
flask
flask_cors
gunicorn
gevent
